backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
google-genai
PyPDF2 